refactor: use async_to_sync_method
