+++ graphgen/operators/__init__.py
@@ -1,22 +1,13 @@
+from graphgen.operators.build_kg.extract_kg import extract_kg
 from graphgen.operators.generate.generate_cot import generate_cot
-from graphgen.operators.kg.extract_kg import extract_kg
 from graphgen.operators.search.search_all import search_all
 
 from .judge import judge_statement
 from .quiz import quiz
+from .read import read_files
+from .split import split_chunks
 from .traverse_graph import (
     traverse_graph_for_aggregated,
     traverse_graph_for_atomic,
     traverse_graph_for_multi_hop,
 )
-
-__all__ = [
-    "extract_kg",
-    "quiz",
-    "judge_statement",
-    "search_all",
-    "traverse_graph_for_aggregated",
-    "traverse_graph_for_atomic",
-    "traverse_graph_for_multi_hop",
-    "generate_cot",
-]
