+++ graphgen/operators/read/__init__.py
@@ -1,0 +1,1 @@
+from .read_files import read_files
