+++ graphgen/operators/split/__init__.py
@@ -1,0 +1,1 @@
+from .split_chunks import split_chunks
