+++ graphgen/utils/__init__.py
@@ -13,3 +13,4 @@ from .hash import compute_args_hash, compute_content_hash
 from .help_nltk import NLTKHelper
 from .log import logger, parse_log, set_logger
 from .loop import create_event_loop
+from .wrap import async_to_sync_method
